add: login_required decorators for profile views and add,edit and delete product views to ensure site security (only site admin can make changes to store

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, get_object_or_404, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 
 from .models import Record, Label
@@ -57,10 +58,15 @@ def product_detail(request, record_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     '''
     Allow superuser/admin to add a product to the store
     '''
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store Admin can do that.')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -82,10 +88,15 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, record_id):
     '''
     Edit a record in the store
     '''
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store Admin can do that.')
+        return redirect(reverse('home'))
 
     record = get_object_or_404(Record, pk=record_id)
     if request.method == 'POST':
@@ -110,10 +121,15 @@ def edit_product(request, record_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, record_id):
     '''
     Delete a record from the store
     '''
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store Admin can do that.')
+        return redirect(reverse('home'))
 
     record = get_object_or_404(Record, pk=record_id)
     record.delete()
